+++ src/EventSerializers/JsonEventSerializer.php
@@ -4,8 +4,6 @@ namespace Spatie\EventSourcing\EventSerializers;
 
 use Spatie\EventSourcing\StoredEvents\ShouldBeStored;
 use Symfony\Component\Serializer\Encoder\JsonEncoder;
-use Symfony\Component\Serializer\Normalizer\DateTimeNormalizer;
-use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
 use Symfony\Component\Serializer\Serializer as SymfonySerializer;
 
 class JsonEventSerializer implements EventSerializer
@@ -15,7 +13,10 @@ class JsonEventSerializer implements EventSerializer
     public function __construct()
     {
         $encoders = [new JsonEncoder()];
-        $normalizers = [new DateTimeNormalizer(), new ObjectNormalizer()];
+        $normalizers = array_map(
+            fn ($className) => new $className,
+            config('event-sourcing.event_normalizers')
+        );
 
         $this->serializer = new SymfonySerializer($normalizers, $encoders);
     }
@@ -23,8 +24,7 @@ class JsonEventSerializer implements EventSerializer
     public function serialize(ShouldBeStored $event): string
     {
         /*
-         * We call __sleep so `Illuminate\Queue\SerializesModels` will
-         * prepare all models in the event for serialization.
+         * We call __sleep so that the event can prepare for serialization.
          */
         if (method_exists($event, '__sleep')) {
             $event->__sleep();
@@ -35,12 +35,6 @@ class JsonEventSerializer implements EventSerializer
 
     public function deserialize(string $eventClass, string $json, string $metadata = null): ShouldBeStored
     {
-        $restoredEvent = $this->serializer->deserialize($json, $eventClass, 'json');
-
-        /*
-         *  We call manually serialize and unserialize to trigger
-         * `Illuminate\Queue\SerializesModels` model restoring capabilities.
-         */
-        return unserialize(serialize($restoredEvent));
+        return $this->serializer->deserialize($json, $eventClass, 'json');
     }
 }