+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 All notable changes to `laravel-event-sourcing` will be documented in this file:
 
+## 4.5.3 - 2020-10-02
+
+- re-use existing instance of `ShouldBeStored` when possible (#158)
+
+## 4.5.2 - 2020-10-02
+
+- fix Paths and Reference URL in event-sourcing.php config file (#159)
+
+## 4.5.1 - 2020-09-27
+
+- added `loadUuid` (#156)
+
+## 4.5.0 - 2020-09-25
+
+- make normalizers configurable (#153)
+
+## 4.4.0 - 2020-09-18
+
+- Support `then` for aggregate root fakes (#154)
+
+## 4.3.1 - 2020-09-09
+
+- Support Laravel 8
+
+## 4.3.0 - 2020-08-24
+
+- support Carbon dates in events (#137)
+
 ## 4.2.0 - 2020-08-06
 
 - allow events to be dispatched from an aggregate root (#135)